Causal Effect Functions

These functions create individual level causal effects, tau_i, that can be combined with a potential outcome to control to create a potential outcome to treatment.

Usage

tau_norm(ybase, tau_sds, covariate)

tau_norm_outliers(ybase, tau_sds, covariate)

tau_norm_covariate(ybase, tau_sds, covariate)

tau_norm_covariate_outliers(ybase, tau_sds, covariate)

tau_norm_covariate_cont(ybase, tau_sds, covariate)

tau_norm_covariate_levels(ybase, tau_sds, covariate)

Arguments

ybase

Is a vector of potential outcome to control

tau_sds

is the number of sds of ybase to shift the distribution

covariate

Contains information about covariates currently a character name of a column in idat. It is NULL if not used. Mostly it is a vector the same length as ybase.

Value

A vector of individual level causal effects (taus) that we will add to ybase (potential outcome to control) to get y1var or potential outcome to treatment.

Functions

tau_norm(): A basic function with no outliers

tau_norm_outliers(): Draws from a Normal but also adds a few outliers.

tau_norm_covariate(): A basic function that specifies a tau_sds*2 size effect if cov>median(cov) and otherwise is a tau_sds/2 size effect. The idea is to keep the average individual effect size the same as other functions — i.e. about tau_sds — but to make a strong but simple relationship with a covariate.

tau_norm_covariate_outliers(): A basic function that specifies a tau_sds size effect if cov>median(cov) and otherwise is a tau_sds/4 size effect

tau_norm_covariate_cont(): A basic function that specifies a tau_sds size effect if cov>median(cov) and otherwise is a tau_sds/4 size effect

tau_norm_covariate_levels(): A basic function that specifies a tau_sds size effect that varies, randomly, by level of covariate. We envision this to be used with covariate that have relatively few levels into which the outcome values can be grouped.