Example tree-testing data

Synthetic data used in examples and tests for manytestsr.

Usage

idt

bdt1

Format

idt

Individual-level data.table with outcomes and assignments.

bdt1

Block-level data.table containing summaries per block.

An object of class data.table (inherits from data.frame) with 64 rows and 9 columns.

Source

See data-raw/make_k_ary_tree_data.R; produced with TreeTestSim (≥ X.Y.Z) and a fixed RNG seed on 2025-11-08.

Details

Generated once and stored as package data to avoid slow/regenerative simulation in examples/tests. The full recipe lives in data-raw/make_k_ary_tree_data.R.

References

Bowers et al., TreeTestSim, manytestsr.

Examples

str(idt)
#> Classes ‘data.table’ and 'data.frame':	3200 obs. of  13 variables:
#>  $ bF          : Factor w/ 64 levels "22","23","24",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ b           : int  22 22 22 22 22 22 22 22 22 22 ...
#>  $ id          : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ bary0       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ y0          : num  -0.966 2.067 -2.617 -0.452 1.039 ...
#>  $ nonnull     : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ lvls_fac    : Factor w/ 64 levels "2.6.22","2.6.23",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ y1          : num  -0.966 2.067 -2.617 -0.452 1.039 ...
#>  $ trt         : num  1 0 0 0 1 0 1 1 0 1 ...
#>  $ Y           : num  -0.966 2.067 -2.617 -0.452 1.039 ...
#>  $ trtF        : Factor w/ 2 levels "0","1": 2 1 1 1 2 1 2 2 1 2 ...
#>  $ y1_half_tau1: num  0.492 3.308 -1.329 0.46 1.963 ...
#>  $ Y_half_tau1 : num  0.492 2.067 -2.617 -0.452 1.963 ...
#>  - attr(*, ".internal.selfref")=<externalptr> 
#>  - attr(*, "sorted")= chr "bF"
#>  - attr(*, "index")= int(0) 
#>   ..- attr(*, "__nonnull")= int [1:3200] 201 202 203 204 205 206 207 208 209 210 ...
str(bdt1)
#> Classes ‘data.table’ and 'data.frame':	64 obs. of  9 variables:
#>  $ node    : int  22 23 24 25 26 27 28 29 30 31 ...
#>  $ level   : int  3 3 3 3 3 3 3 3 3 3 ...
#>  $ parent  : num  6 6 6 6 7 7 7 7 8 8 ...
#>  $ nonnull : logi  TRUE TRUE TRUE TRUE FALSE FALSE ...
#>  $ leaf    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ lvls_fac: Factor w/ 64 levels "2.6.22","2.6.23",..: 1 2 3 4 5 6 7 8 9 10 ...
#>  $ nb      : num  50 50 50 50 50 50 50 50 50 50 ...
#>  $ bary0   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ bF      : Factor w/ 64 levels "22","23","24",..: 1 2 3 4 5 6 7 8 9 10 ...
#>  - attr(*, ".internal.selfref")=<externalptr>