Make a plot of the nodes

Given the results of the splitting and testing algorithm in the form of a graph from make_results_tree, make a node level data set for use in reporting results in terms of a binary tree graph. This does not print or plot the graph. You'll need to do that with the resulting object.

Usage

make_results_ggraph(res_graph)

Arguments

res_graph

A tidygraph object produced from make_results_tree

Value

A ggraph object